update cpp accelarator

+++ jiojio/jiojio_cpp/compiler.sh
@@ -1,2 +1,10 @@
-rm -rf build/*
+#!/bin/bash
+
+if [ ! -d build ]; then
+    mkdir build
+else
+    rm -rf build/*
+fi
 cd build && cmake .. && make
+
+cd .. && python test_cpp.py
